requirements.txt
altgraph==0.17.3
distro==1.8.0
et-xmlfile==1.1.0
fsspec==2023.5.0
install==1.3.5
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
pefile==2023.2.7
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
six==1.16.0
tabula-py==2.7.0
tk==0.1.0
tzdata==2023.3
